requirements.txt
drfp==0.3.2
jax==0.4.8
numpy==1.23.4
numpyro==0.11.0
pandas==1.5.1
pyro_ppl==1.8.4
scikit_learn==1.1.3
scipy==1.9.1
torch==1.13.1+cu117
